moved all automod parsing code to AutoModMessageParsing
