+++ app/src/main/java/com/example/clicker/network/domain/TwitchEventSubscriptionWebSocket.kt
@@ -2,8 +2,9 @@ package com.example.clicker.network.domain
 
 import androidx.compose.runtime.State
 import com.example.clicker.network.models.twitchStream.ChatSettingsData
-import com.example.clicker.network.websockets.AutoModMessageUpdate
-import com.example.clicker.network.websockets.AutoModQueueMessage
+import com.example.clicker.network.repository.util.AutoModMessageUpdate
+import com.example.clicker.network.repository.util.AutoModQueueMessage
+
 import com.example.clicker.presentation.modView.ModActionData
 import kotlinx.coroutines.flow.StateFlow
