+++ app/src/main/java/com/example/clicker/network/websockets/TwitchEventSubWebSocket.kt
@@ -6,6 +6,9 @@ import androidx.annotation.RequiresApi
 import com.example.clicker.R
 import com.example.clicker.network.domain.TwitchEventSubscriptionWebSocket
 import com.example.clicker.network.models.twitchStream.ChatSettingsData
+import com.example.clicker.network.repository.util.AutoModMessageParsing
+import com.example.clicker.network.repository.util.AutoModMessageUpdate
+import com.example.clicker.network.repository.util.AutoModQueueMessage
 import com.example.clicker.network.repository.util.ChatSettingsParsing
 import com.example.clicker.network.repository.util.ModActionParsing
 import com.example.clicker.presentation.modView.ModActionData
@@ -27,7 +30,8 @@ import javax.inject.Inject
 
 class TwitchEventSubWebSocket @Inject constructor(
     private val modActionParsing:ModActionParsing,
-    private val channelSettingsParsing: ChatSettingsParsing
+    private val channelSettingsParsing: ChatSettingsParsing,
+    private val autoModMessageParsing: AutoModMessageParsing
 ): TwitchEventSubscriptionWebSocket, WebSocketListener() {
     private var client: OkHttpClient = OkHttpClient.Builder().build()
     var webSocket: WebSocket? = null
@@ -74,14 +78,20 @@ class TwitchEventSubWebSocket @Inject constructor(
         }
 
         if(notificationTypeIsNotification(text)){
-            //Log.d("TwitchEventSubWebSocket","notificationTypeIsNotification  ->${parseAutoModQueueMessage(text)}")
+
             val subscriptionType = parseSubscriptionType(text)
 
             //todo: These could probably be stored in maps
             when (subscriptionType) {
                 "automod.message.hold" -> {
-                    _autoModMessageQueue.tryEmit(parseAutoModQueueMessage(text))
+                    _autoModMessageQueue.tryEmit(autoModMessageParsing.parseAutoModQueueMessage(text))
                 }
+                "automod.message.update" -> {
+                    val messageId = autoModMessageParsing.parseMessageId(text) ?: ""
+                    val messageUpdate = autoModMessageParsing.checkUpdateStatus(text, messageId)
+                    _messageIdForAutoModQueue.tryEmit(messageUpdate)
+                }
+
                 "channel.moderate" -> {
                     Log.d("ChannelModerateParsing", "TIME TO PARSE!!!!")
                     val action = modActionParsing.parseActionFromString(text)
@@ -92,11 +102,7 @@ class TwitchEventSubWebSocket @Inject constructor(
                     )
 
                 }
-                "automod.message.update" -> {
-                    val messageId = parseMessageId(text) ?: ""
-                    val messageUpdate = checkUpdateStatus(text, messageId)
-                    _messageIdForAutoModQueue.tryEmit(messageUpdate)
-                }
+
                 "channel.chat_settings.update" -> {
                     Log.d("ChatSettingsParsing",text)
                     val parsedChatSettingsData = channelSettingsParsing.parseChatSettingsData(text)
@@ -108,23 +114,6 @@ class TwitchEventSubWebSocket @Inject constructor(
         Log.d("TwitchEventSubWebSocket","onMessage() text ->$text")
         Log.d("createAnotherSubscriptionEvent","onMessage text -->$text")
 
-    }
-    fun checkUpdateStatus(
-        text:String,
-        messageId: String
-    ):AutoModMessageUpdate{
-        val type =parseStatusType(text)
-        if(type =="denied"){
-            return AutoModMessageUpdate(
-                approved = false,
-                messageId = messageId
-            )
-        }else{
-            return AutoModMessageUpdate(
-                approved = true,
-                messageId = messageId
-            )
-        }
     }
 
     override fun onClosing(webSocket: WebSocket, code: Int, reason: String) {
@@ -237,20 +226,7 @@ fun parseSubscriptionType(stringToParse:String):String{
     val subscriptionType = messageTypeRegex.find(stringToParse)?.groupValues?.get(1)?.replace("\"","")
     return subscriptionType?:""
 }
-fun parseMessageId(stringToParse:String):String?{
-    val messageIdRegex = "\"message_id\":([^=]+)".toRegex()
-    val allFoundMessageIdList = messageIdRegex.findAll(stringToParse).toList()
-    return if(allFoundMessageIdList.size >=2){
-        val messageIdNoQuotes=allFoundMessageIdList[1].groupValues[1].replace("\"","")
-        val newMessageRegex = "[^,]+".toRegex()
-        val desiredMessageId = newMessageRegex.find(messageIdNoQuotes)?.value
-        desiredMessageId
-    }else{
-        null
-    }
 
-
-}
 
 fun parseStatusType(stringToParse:String):String?{
     val messageIdRegex = "\"status\":([^,]+)".toRegex()
@@ -270,66 +246,3 @@ fun parseStatusType(stringToParse:String):String?{
 
 
 
-/**
- * parseAutoModQueueMessage is a function meant to parse out the username, category and fullText from the
- * message that is sent from the Twitch servers.
- *
- * @param stringToParse represents what was sent from the twitch servers
- * @return [AutoModQueueMessage] object
- * */
-fun parseAutoModQueueMessage(stringToParse:String):AutoModQueueMessage{
-    val usernameRegex = "\"user_name\":([^,]+)".toRegex()
-    val textRegex = "\"text\":([^,]+)".toRegex()
-    val categoryRegex = "\"category\":([^,]+)".toRegex()
-    val userIdRegex = "\"user_id\":([^,]+)".toRegex()
-
-    //extra parsing because there is multiple `message_id`
-    val messageIdRegex = "\"message_id\":([^=]+)".toRegex()
-    val allFoundMessageIdList = messageIdRegex.findAll(stringToParse).toList()
-    val messageIdNoQuotes=allFoundMessageIdList[1].groupValues[1].replace("\"","")
-    val newMessageRegex = "[^,]+".toRegex()
-    val desiredMessageId = newMessageRegex.find(messageIdNoQuotes)?.value
-
-
-    val username = usernameRegex.find(stringToParse)?.groupValues?.get(1)?.replace("\"","")
-    val fullText = textRegex.find(stringToParse)?.groupValues?.get(1)?.replace("\"","")
-    val category = categoryRegex.find(stringToParse)?.groupValues?.get(1)?.replace("\"","")
-    val userId = userIdRegex.find(stringToParse)?.groupValues?.get(1)?.replace("\"","")
-    val messageId = desiredMessageId
-
-    return AutoModQueueMessage(
-        username = username ?:"",
-        fullText = fullText ?:"",
-        category = category ?:"",
-        userId = userId ?:"",
-        messageId = messageId ?:""
-    )
-}
-
-
-
-
-/**
- * AutoModQueueMessage is a data class that will represent the parsed data send from
- * [Notification message Websocket](https://dev.twitch.tv/docs/eventsub/handling-websocket-events/#notification-message)
- *
- * @param username represents the user that sent the [fullText] message
- * @param fullText represents the message that the user sent
- * @param category represents the category that the user's [fullText] message falls under, ie swearing
- * @param messageId represents the unique ID of the user's flagged message
- * @param userId represents the unique ID of the user sending the message
- * */
-data class AutoModQueueMessage(
-    val username:String ="",
-    val fullText:String ="",
-    val category: String = "",
-    val messageId:String,
-    val userId:String,
-    var approved:Boolean? = null,
-    var swiped:Boolean = false,
-)
-
-data class AutoModMessageUpdate(
-    val approved: Boolean,
-    val messageId: String
-)
