+++ app/src/main/java/com/example/clicker/presentation/stream/views/streamManager/ModView.kt
@@ -72,7 +72,7 @@ import com.example.clicker.R
 import com.example.clicker.network.clients.BlockedTerm
 import com.example.clicker.network.models.websockets.TwitchUserData
 import com.example.clicker.network.repository.EmoteListMap
-import com.example.clicker.network.websockets.AutoModQueueMessage
+
 import com.example.clicker.presentation.modView.ListTitleValue
 import com.example.clicker.presentation.modView.ModActionData
 import com.example.clicker.presentation.modView.ModViewDragStateViewModel
